feat: se agrega programación defensiva

+++ src/main/java/com/ronal/educlass/service/impl/AlumnoServiceImpl.java
@@ -4,12 +4,16 @@ import com.ronal.educlass.entity.Alumno;
 import com.ronal.educlass.repository.IAlumnoRepository;
 import com.ronal.educlass.service.IAlumnoService;
 import com.ronal.educlass.utils.AlumnoUtils;
+import lombok.extern.slf4j.Slf4j;
+import org.ronal.exceptionr.custom.BusinessException;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
 @Service
+@Slf4j
 public class AlumnoServiceImpl implements IAlumnoService {
     private final IAlumnoRepository alumnoRepository;
 
@@ -20,8 +24,9 @@ public class AlumnoServiceImpl implements IAlumnoService {
     @Override
     public Alumno crearAlumno(Alumno alumno) {
         Optional<Alumno> alumnoExistente = this.alumnoRepository.findByDni(alumno.getDni());
-        if (alumnoExistente.isPresent()){
-            // TODO: lanzar excepción
+        if (alumnoExistente.isPresent()) {
+            throw new BusinessException("ALUMNO_EXISTE", HttpStatus.INTERNAL_SERVER_ERROR,
+                    "El Alumno con el DNI: " + alumno.getDni() + " ya se encuentra registrado en el Sistema");
         }
         String codAlumno = AlumnoUtils.generarCodAlumno(alumno);
         String correoAlumno = AlumnoUtils.generarCorreoAlumno(codAlumno);
@@ -39,15 +44,18 @@ public class AlumnoServiceImpl implements IAlumnoService {
     @Override
     public void eliminarAlumno(Long idAlumno) {
         Alumno alumnoEliminar = this.buscarAlumnoPorId(idAlumno);
+        if (alumnoEliminar == null) {
+            throw new BusinessException("ALUMNO_NO_ENCONTRADO", HttpStatus.NOT_FOUND,
+                    "El alumno que desea eliminar no existe en el sistema");
+        }
         this.alumnoRepository.delete(alumnoEliminar);
     }
 
     @Override
     public Alumno actualizarAlumno(Alumno alumno) {
-
         Alumno alumnoEncontrado = this.buscarAlumnoPorId(alumno.getId());
         Alumno alumnoActualizado = null;
-        if (alumnoEncontrado != null){
+        if (alumnoEncontrado != null) {
             alumnoActualizado = Alumno.builder()
                     .nombre(alumno.getNombre())
                     .apellido1(alumno.getApellido1())
@@ -59,8 +67,10 @@ public class AlumnoServiceImpl implements IAlumnoService {
                     .codigo(alumno.getCodigo())
                     .build();
             this.alumnoRepository.save(alumno);
+            return alumnoActualizado;
         }
-        return alumnoActualizado;
+        throw new BusinessException("ALUMNO_NO_ENCONTRADO", HttpStatus.NOT_FOUND,
+                "El alumno que desea actualizar no existe en el sistema");
     }
 
     @Override
